desktop_app/requirements.txt
PySide6==6.9.0
httpx==0.27.2
python-dotenv==1.1.0
pydantic==2.9.2
